+++ template_engines/backends/odt.py
@@ -4,7 +4,8 @@ from django.template import Context
 from django.template.context import make_context
 
 from .abstract import AbstractTemplate, ZipAbstractEngine
-from .utils import modify_libreoffice_doc
+from .utils import modify_content_document
+from .utils_odt import add_image_in_odt_template
 
 
 class OdtTemplate(AbstractTemplate):
@@ -96,7 +97,7 @@ class OdtTemplate(AbstractTemplate):
         style_list_properties = soup.new_tag('style:list-level-properties')
         style_list_properties.attrs = {"text:list-level-position-and-space-mode": "label-alignment"}
         style_number.append(style_list_properties)
-        # Space between text and number : "listtab nothin or space possible
+        # Space between text and number : "listtab, nothing or space allowed
         style_alignment = soup.new_tag('style:list-level-label-alignment')
         style_alignment.attrs = {"text:label-followed-by": "space", "fo:text-indent": "0.435cm"}
         style_list_properties.append(style_alignment)
@@ -117,7 +118,7 @@ class OdtTemplate(AbstractTemplate):
         style_list_properties = soup.new_tag('style:list-level-properties')
         style_list_properties.attrs = {"text:list-level-position-and-space-mode": "label-alignment"}
         style_number.append(style_list_properties)
-        # Space between text and number : "listtab nothin or space possible
+        # Space between text and number : "listtab, nothing or space allowed
         style_alignment = soup.new_tag('style:list-level-label-alignment')
         style_alignment.attrs = {"text:label-followed-by": "space", "fo:text-indent": "0.635cm"}
         style_list_properties.append(style_alignment)
@@ -170,7 +171,9 @@ class OdtTemplate(AbstractTemplate):
         soup = BeautifulSoup(rendered, features='xml')
         soup = self.clean(soup)
         soup = self.replace_inputs(soup)
-        odt_content = modify_libreoffice_doc(self.template_path, 'content.xml', str(soup))
+        odt_content = modify_content_document(self.template_path, 'content.xml', str(soup))
+        for key, image in context.get('images', {}).items():
+            odt_content = add_image_in_odt_template(odt_content, image, key)
         return odt_content