+++ template_engines/templatetags/utils.py
@@ -3,6 +3,9 @@ from io import BytesIO
 
 from PIL import Image
 
+from django.template.base import FilterExpression, kwarg_re
+
+
 DIM_REGEX = r'^(?P<v>(\d|\.)+)(?P<u>[a-z]*)$'
 
 DOCX_PAGE_WIDTH = 6120130
@@ -86,3 +89,43 @@ def resize(bimage, width, height, odt=True):
         height = size_parser(height, odt=odt)
 
     return width, height
+
+
+def parse_tag(token, parser):
+    """
+    Generic template tag parser.
+
+    Returns a three-tuple: (tag_name, args, kwargs)
+
+    tag_name is a string, the name of the tag.
+
+    args is a list of FilterExpressions, from all the arguments that didn't look like kwargs,
+    in the order they occurred, including any that were mingled amongst kwargs.
+
+    kwargs is a dictionary mapping kwarg names to FilterExpressions, for all the arguments that
+    looked like kwargs, including any that were mingled amongst args.
+
+    (At rendering time, a FilterExpression f can be evaluated by calling f.resolve(context).)
+    """
+    # Split the tag content into words, respecting quoted strings.
+    bits = token.split_contents()
+
+    # Pull out the tag name.
+    tag_name = bits.pop(0)
+
+    # Parse the rest of the args, and build FilterExpressions from them so that
+    # we can evaluate them later.
+    args = []
+    kwargs = {}
+    for bit in bits:
+        bit = bit.replace('&quot;', '"')
+        # Is this a kwarg or an arg?
+        match = kwarg_re.match(bit)
+        kwarg_format = match and match.group(1)
+        if kwarg_format:
+            key, value = match.groups()
+            kwargs[key] = FilterExpression(value, parser)
+        else:
+            args.append(FilterExpression(bit, parser))
+
+    return tag_name, args, kwargs