+++ template_engines/tests/test_template_tags/test_odt_tags.py
@@ -1,9 +1,14 @@
+import io
 from bs4 import BeautifulSoup
+from requests.exceptions import ConnectionError
 from unittest import mock
 
+from django.template import Context, Template
+from django.template.exceptions import TemplateSyntaxError
 from django.test import TestCase
 
 from template_engines.templatetags import odt_tags
+from template_engines.tests.settings import IMAGE_PATH
 
 
 class FilterFromHTMLTestCase(TestCase):
@@ -64,3 +69,113 @@ class FilterFromHTMLTestCase(TestCase):
         html = '<p><br></p>'
         soup = odt_tags.from_html(html)
         self.assertEqual('<text:p><text:line-break/></text:p>', str(soup))
+
+
+@mock.patch('secrets.token_hex', return_value='test')
+class ImageLoaderTestCase(TestCase):
+    def test_image_loader_object(self, token):
+        context = Context({'image': {'content': open(IMAGE_PATH, 'rb').read()}})
+        template_to_render = Template('{% load odt_tags %}{% image_loader image %}{% image_loader image %}')
+        rendered_template = template_to_render.render(context)
+        self.assertEqual(rendered_template.count('<draw:frame draw:name="{name}"'.format(name=token.return_value)), 2)
+
+    def test_image_loader_resize(self, token):
+        context = Context({'image': {'content': open(IMAGE_PATH, 'rb').read()}})
+        template_to_render = Template('{% load odt_tags %}{% image_loader image width="32" height="42" %}')
+        rendered_template = template_to_render.render(context)
+        self.assertNotIn('svg:width="16697.0" svg:height="5763.431472081218"', rendered_template)
+        self.assertIn('svg:width="32.0" svg:height="42.0"', rendered_template)
+
+    def test_image_url_loader_fail(self, token):
+        with self.assertRaises(TemplateSyntaxError) as cm:
+            Template('{% load odt_tags %}{% image_loader image=image %}')
+        self.assertEqual('Usage: {% image_loader [image] width="5000" height="5000" %}', str(cm.exception))
+
+
+@mock.patch('secrets.token_hex', return_value='test')
+class ImageUrlLoaderTestCase(TestCase):
+    @mock.patch('requests.get')
+    def test_image_url_loader_object(self, mocked_get, token):
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.content = open(IMAGE_PATH, 'rb').read()
+        context = Context({'url': "https://test.com"})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader url %}')
+
+        rendered_template = template_to_render.render(context)
+        self.assertEqual('<draw:frame draw:name="{name}" svg:width="16697.0" svg:height="5763.431472081218" '
+                         'text:anchor-type="paragraph" draw:z-index="0" text:anchor-type="paragraph">'
+                         '<draw:image xlink:href="Pictures/{name}" xlink:show="embed" xlink:actuate="onLoad">'
+                         '</draw:image></draw:frame>'.format(name=token.return_value), rendered_template)
+
+    @mock.patch('requests.get')
+    def test_image_url_loader_url(self, mocked_get, token):
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.content = open(IMAGE_PATH, 'rb').read()
+        context = Context({})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader "https://test.com" %}')
+
+        rendered_template = template_to_render.render(context)
+        self.assertEqual('<draw:frame draw:name="{name}" svg:width="16697.0" svg:height="5763.431472081218" '
+                         'text:anchor-type="paragraph" draw:z-index="0" text:anchor-type="paragraph">'
+                         '<draw:image xlink:href="Pictures/{name}" xlink:show="embed" xlink:actuate="onLoad">'
+                         '</draw:image></draw:frame>'.format(name=token.return_value), rendered_template)
+
+    @mock.patch('requests.get')
+    def test_image_url_loader_resize(self, mocked_get, token):
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.content = open(IMAGE_PATH, 'rb').read()
+        context = Context({'url': "https://test.com"})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader url width="32" height="42" %}')
+        rendered_template = template_to_render.render(context)
+        self.assertNotIn('svg:width="16697.0" svg:height="5763.431472081218"', rendered_template)
+        self.assertIn('svg:width="32.0" svg:height="42.0"', rendered_template)
+
+    @mock.patch('requests.get')
+    def test_image_url_loader_fail(self, mocked_get, token):
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.content = open(IMAGE_PATH, 'rb').read()
+        with self.assertRaises(TemplateSyntaxError) as cm:
+            Template('{% load odt_tags %}{% image_url_loader url="https://test.com" %}')
+        self.assertEqual('Usage: {% image_url_loader [url] width="5000" '
+                         'height="5000" request="GET" data="{"data": "example"}"%}', str(cm.exception))
+
+    @mock.patch('requests.get')
+    @mock.patch('sys.stderr', new_callable=io.StringIO)
+    def test_image_url_loader_picture_not_accessible(self, mock_out, mocked_get, token):
+        mocked_get.return_value.status_code = 404
+        mocked_get.return_value.content = b''
+        context = Context({})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader "https://test.com" %}')
+        template_to_render.render(context)
+        self.assertEqual(mock_out.getvalue(), 'The picture is not accessible (Error: 404)\n')
+
+    @mock.patch('requests.get')
+    @mock.patch('sys.stderr', new_callable=io.StringIO)
+    def test_image_url_loader_picture_connection_error(self, mock_out, mocked_get, token):
+        mocked_get.side_effect = ConnectionError
+        context = Context({})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader "https://test.com" %}')
+        template_to_render.render(context)
+        self.assertEqual(mock_out.getvalue(), 'Connection Error, check the url given\n')
+
+    @mock.patch('requests.get')
+    @mock.patch('sys.stderr', new_callable=io.StringIO)
+    def test_image_url_loader_picture_wrong_request(self, mock_out, mocked_get, token):
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.content = open(IMAGE_PATH, 'rb').read()
+        context = Context({})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader "https://test.com" request="WRONG" %}')
+        template_to_render.render(context)
+        self.assertEqual(mock_out.getvalue(), 'Type of request specified not allowed\n')
+
+    @mock.patch('requests.post')
+    def test_image_url_loader_picture_post_request(self, mocked_post, token):
+        mocked_post.return_value.status_code = 200
+        mocked_post.return_value.content = open(IMAGE_PATH, 'rb').read()
+        context = Context({})
+        template_to_render = Template('{% load odt_tags %}{% image_url_loader "https://test.com" request="POST" %}')
+        rendered_template = template_to_render.render(context)
+        self.assertEqual('<draw:frame draw:name="{name}" svg:width="16697.0" svg:height="5763.431472081218" '
+                         'text:anchor-type="paragraph" draw:z-index="0" text:anchor-type="paragraph">'
+                         '<draw:image xlink:href="Pictures/{name}" xlink:show="embed" xlink:actuate="onLoad">'
+                         '</draw:image></draw:frame>'.format(name=token.return_value), rendered_template)